+++ HISTORY.rst
@@ -6,6 +6,8 @@ Release History
 * New features for BaseGridPoints, namely index() which returns the index at which that BaseGridpoint exists in the container, and distance() which calculates the distance (in points) from one BaseGridPoint to another.
 * New linker functions for returning remote linkers via conencted Summits or Saddles
 * SpotElevationContainer object now support index()
+* GridPointContainer now supports highest() and lowest()
+* GridPoint now Hashable
 
 0.5.10 (September 17, 2018)
 +++++++++++++++++++++++++++